Basic solution
This solution uses a simple in memory cache provider as no caching
library is available on the classpath.

Caching is opt-in and enabled via the `@EnableCaching` annotation. Once
this is done, method needs to be marked to add a "cache aspect" to it.

In this example, `findByCode` is cached using the cache named
"countries". As there are no further customization, the `code` parameter
is used to compute the key of the entry in the cache.

As we are using a basic in-memory cache provider, caches will be created
on-the-fly. When you switch to a real cache provider such as ehcache,
you will need to configure the caches of your application using the
configuration infrastructure of the cache library you've chosen.